wgpu: Split image Memory and Pipeline in separate module files. Do not panic on ImageError
